+++ static/D3_js/interaction/canvas_move.js
