+++ static/D3_js/interaction/info_table.js
