--- static/test copy.html
@@ -1,205 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="https://neo4j-documentation.github.io/developer-resources/language-guides/assets/css/main.css">
-    <title>Neo4j RCN</title>
-</head>
-
-<body>
-<div id="graph">
-</div>
-
-<!-- Load d3.js -->
-<script src="https://cdnjs.cloudflare.com/ajax/libs/d3/7.8.4/d3.min.js"></script>
-<script>
-
-var width = 4000;
-var height = 2000;
-var nodes = [
-  { id: "A" },
-  { id: "B" },
-  { id: "C" },
-  { id: "D" }
-];
-var links = [
-  { source: "A", target: "B" },
-  { source: "B", target: "C" },
-  { source: "C", target: "D" },
-  { source: "D", target: "A" }
-];
-
-var svg =
-  d3.select("#graph").append("svg").attr("width", width).attr("height", height)
-    // .append("g")
-    // .attr("transform", "translate(" + (width / 2) + "," + (height / 2) + ")")
-    ;
-
-var clicked = {};
-var last_clicked_node;
-
-var simulation = d3.forceSimulation(nodes)
-  .force("link", d3.forceLink(links).id(function(d) { return d.id; }))
-  .force("charge", d3.forceManyBody().strength(-5))
-  .force("center", d3.forceCenter(300, 200))
-  .force("collision", d3.forceCollide().radius(function(d) { return 20; }));
-
-// create the links
-var link = svg.selectAll("line")
-  .data(links)
-  .enter()
-  .append("line")
-  .attr("stroke", "black")
-  .attr("stroke-width", 2);
-
-// create the nodes
-var node = svg.selectAll("circle")
-  .data(nodes)
-  .enter()
-  .append("circle")
-  .attr("r", 10)
-  .attr("fill", "green")
-  .attr("stroke", "light-grey")
-//   .on("click", function(event, d) {
-//     addOrRemoveTooltip(event, d);
-//     })
-    .attr("cursor", "pointer");;
-
-
-// const node = svg.append("circle")
-//     .attr("cx", 10)
-//     .attr("cy", 10)
-//     .attr("r", 45)
-//     // .attr("transform", `translate(10,10) scale(45)`)
-//     .attr("fill", "green")
-//     .attr("stroke", "light-grey")
-// //   .on("click", function(event, d) {
-// //     addOrRemoveTooltip(event, d);
-// //     })
-//     .attr("cursor", "pointer");
-
-addOrRemoveTooltip(node);
-
-simulation.on("tick", function() {
-  link
-    .attr("x1", function(d) { return d.source.x; })
-    .attr("y1", function(d) { return d.source.y; })
-    .attr("x2", function(d) { return d.target.x; })
-    .attr("y2", function(d) { return d.target.y; });
-
-  node
-    .attr("cx", function(d) { return d.x; })
-    .attr("cy", function(d) { return d.y; });
-});
-
-var arcs = svg.selectAll("arcs")
-            .data([
-                { start: 0, end: 1/3, label: "hide",icon: "../images/remove_icon.png" },
-                { start: 1/3, end: 2/3, label: "link",  icon: "../images/arrows_expand_icon.png" },
-                { start: 2/3, end: 1, label: "lock", icon:["../images/lock_closed_icon.png", "../images/lock_open_icon.png"] }
-            ])
-            .enter();
-
-function addOrRemoveTooltip(selection) {
-
-    selection.on("click", function(event, d){
-    if (clicked[d.id]) {
-        arcs.selectAll("path").remove();
-        arcs.selectAll("image").remove();
-        clicked[d.id] = false;
-    } 
-
-    else {
-        if (last_clicked_node && last_clicked_node != d.id) {
-            clicked[last_clicked_node] = false;
-            arcs.selectAll("path").remove();
-            arcs.selectAll("image").remove();
-        }
-        var arc = d3.arc().innerRadius(11).outerRadius(26);
-        
-        // var nodePos = d3.select(this).attr("transform").replace("translate(", "").replace(")", "").split(",");
-        var node_x = d.x;
-        var node_y = d.y;
-
-        
-        //                 // .attr("transform", d => "rotate(" + ((d.start + d.end) * 180 / Math.PI) + ")")
-        //                 ;
-
-        var icons = arcs.selectAll("image")
-                        .data(arcs.data())
-                        .enter().append("image")
-                        .attr("xlink:href", function (d) {
-                             if (d.label == 'lock'){
-                                return d.icon[0];
-                             }
-                             else{
-                                return d.icon;
-                             } })
-                        .attr("width", 18)
-                        .attr("height", 18)
-                        .attr("x", d=>node_x-10+ 18 * Math.cos((d.start + d.end-1/2) * Math.PI))
-                        .attr("y", d=>node_y-10+ 18 * Math.sin((d.start + d.end-1/2) * Math.PI))
-
-
-        arcs.append("path")
-        .attr("d", d => arc({
-                "startAngle": d.start * 2 * Math.PI + 0.01,
-                "endAngle": d.end * 2 * Math.PI - 0.01
-            }))
-
-        .attr("transform", `translate(${node_x},${node_y})`)
-        .attr("fill", "lightgrey")
-        .style("opacity", 0.5)
-
-        
-
-        .on("mouseover", function(event, d) {
-            // change the style of the button
-            d3.select(this)
-            .style("opacity", 0.2);
-        })
-        .on("mouseout", function(event, d) {
-            // restore the style of the button
-            d3.select(this)
-            .style("opacity", 0.5);
-        })
-        .on("click", function(event, d){
-            
-
-            if (d.label == "hide") {
-                d3.select(this).style("opacity", 0.5);
-                
-            }
-            if (d.label == "lock") {
-                d3.select(this).style("opacity", 0.5);
-            }
-            if (d.label == "link") {
-                d3.select(this).style("opacity", 0.5);
-            }        
-        })
-        
-        ;
-
-        
-
-            // .attr("transform", `translate(100,100)`)
-        // var icons = svg.selectAll("image")
-        //                 .data(arcs.data())
-        //                 .enter().append("image")
-        //                 .attr("xlink:href", d => d.icon)
-        //                 .attr("width", 20)
-        //                 .attr("height", 20)
-        //                 .attr("x", node_x)
-        //                 .attr("y", node_y)
-        //                 .attr("transform", d => "rotate(" + ((d.start + d.end) * 180 / Math.PI) + ")");
-
-    clicked[d.id] = true;
-    last_clicked_node = d.id;
-  }
-})
-}
-</script>
-
-</body>
-</html>
