--- static/test.html
@@ -1,106 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="https://neo4j-documentation.github.io/developer-resources/language-guides/assets/css/main.css">
-    <title>Neo4j RCN</title>
-</head>
-
-<body>
-<div id="graph">
-</div>
-
-<!-- Load d3.js -->
-<script type="text/javascript" src="//code.jquery.com/jquery-1.11.0.min.js"></script>
-<script src="https://d3js.org/d3.v7.min.js"></script>
-<script>
-
-var width = 400;
-var height = 200;
-
-var svg =
-  d3.select("#graph").append("svg").attr("width", width).attr("height", height)
-    .append("g")
-    .attr("transform", "translate(" + (width / 2) + "," + (height / 2) + ")")
-    ;
-
-var clicked = false;
-
-
-const node = svg.append("circle")
-    .attr("cx", 10)
-    .attr("cy", 10)
-    .attr("r", 45)
-    // .attr("transform", `translate(10,10) scale(45)`)
-    .attr("fill", "green")
-    .attr("stroke", "light-grey")
-//   .on("click", function(event, d) {
-//     addOrRemoveTooltip(event, d);
-//     })
-    .attr("cursor", "pointer");
-
-addOrRemoveTooltip(node);
-
-function addOrRemoveTooltip(selection) {
-
-    selection.on("click", function(event, d){
-    if (clicked) {
-        d3.selectAll("path").remove();
-        clicked = false;
-    } 
-    else {
-        var arc = d3.arc().innerRadius(47).outerRadius(80);
-        
-        // var nodePos = d3.select(this).attr("transform").replace("translate(", "").replace(")", "").split(",");
-        // var node_x = parseFloat(nodePos[0]) + parseFloat(d3.select(this).attr("r"));
-        // var node_y = parseFloat(nodePos[1]) + parseFloat(d3.select(this).attr("r"));
-
-        svg.selectAll("arcs")
-            .data([
-                { start: 0, end: 1/3, label: "hide" },
-                { start: 1/3, end: 2/3, label: "link" },
-                { start: 2/3, end: 1, label: "lock" }
-            ])
-            .enter().append("path")
-            .attr("d", d => arc({
-                "startAngle": d.start * 2 * Math.PI + 0.01,
-                "endAngle": d.end * 2 * Math.PI - 0.01
-            }))
-
-            // .attr("transform", `translate(${node_x},${node_y})`)
-            .attr("transform", `translate(10,10)`)
-
-            .attr("fill", "lightgrey")
-            .style("opacity", 0.5)
-
-            .on("mouseover", function(event, d) {
-                // change the style of the button
-                d3.select(this)
-                .style("opacity", 1.0);
-            })
-            .on("mouseout", function(event, d) {
-                // restore the style of the button
-                d3.select(this)
-                .style("opacity", 0.5);
-            })
-            .on("click", function(event, d){
-                if (d.label == "hide") {
-                    d3.select(this).style("opacity", 0.1);
-                }
-                if (d.label == "lock") {
-                    d3.select(this).style("opacity", 0.6);
-                }
-                if (d.label == "link") {
-                    d3.select(this).style("opacity", 0.6);
-                }        
-            });
-
-    clicked = true;
-  }
-})
-}
-</script>
-
-</body>
-</html>
